Fix typo in sub_delay external specification. Patch by Jindrich Makovicka <makovick a gmail d com>
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@20203 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpcodecs/vf_vo.c
@@ -19,7 +19,7 @@ extern ass_track_t* ass_track;
 //===========================================================================//
 
 extern int sub_visibility;
-extern double sub_delay;
+extern float sub_delay;
 
 typedef struct vf_vo_data_s {
     double pts;
